Started a semi-working template
This template is based on Dave Lagattuta's dissertation (sent to me on
6/27/2013). I say that it is semi-working because currently it won't read
my .bib files correctly.  Dave had his program setup to run with the bib
info directly in the .tex file rather than in a .bib file.  Another reason it is
semi-working is that you cannot currently run PDFLaTeX since lines
58-65 specify a hyperlink option that only works on dvips.