performed a centroid uncertainty run based on probabilistic bootstrap sampling
I don't really trust these results due to the low purity of the cluster sample in the
bootstrap draws.  This shows a very high confidence in the galaxy-dm offset but
again I am skeptical of the results.

Note that the north centroid result figure are no good since they were not made
from the customized north density fits file.